# Conflicts:
#	client/src/pages/api/openapi/kb/appKbSearch.ts
#	client/src/pages/kb/components/SelectCsvModal.tsx
#	client/src/pages/kb/detail/components/DataCard.tsx

+++ client/src/pages/api/chat/removeHistory.ts
@@ -1,20 +1,45 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
 import { jsonRes } from '@/service/response';
-import { connectToDatabase, Chat } from '@/service/mongo';
+import { connectToDatabase, Chat, ChatItem } from '@/service/mongo';
 import { authUser } from '@/service/utils/auth';
 
-/* 获取历史记录 */
+type Props = {
+  chatId?: string;
+  appId?: string;
+};
+
+/* clear chat history */
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
   try {
-    const { id } = req.query;
+    const { chatId, appId } = req.query as Props;
     const { userId } = await authUser({ req, authToken: true });
 
     await connectToDatabase();
 
-    await Chat.findOneAndRemove({
-      _id: id,
-      userId
-    });
+    if (chatId) {
+      await Promise.all([
+        Chat.findOneAndRemove({
+          chatId,
+          userId
+        }),
+        ChatItem.deleteMany({
+          userId,
+          chatId
+        })
+      ]);
+    }
+    if (appId) {
+      await Promise.all([
+        Chat.deleteMany({
+          appId,
+          userId
+        }),
+        ChatItem.deleteMany({
+          userId,
+          appId
+        })
+      ]);
+    }
 
     jsonRes(res);
   } catch (err) {
